Fix image asset metadata lookup in Studio
