+++ packages/studio/src/components/Preview.tsx
@@ -12,6 +12,7 @@ import {
 } from '../helpers/checkerboard-background';
 import {LIGHT_TEXT} from '../helpers/colors';
 import type {AssetMetadata} from '../helpers/get-asset-metadata';
+import {getPreviewFileType} from '../helpers/get-preview-file-type';
 import type {Dimensions} from '../helpers/is-current-selected-still';
 import {CheckerboardContext} from '../state/checkerboard';
 import {VERTICAL_SCROLLBAR_CLASSNAME} from './Menu/is-menu-item';
@@ -40,50 +41,6 @@ const assetMetadataErrorContainer: React.CSSProperties = {
 	position: 'absolute',
 	height: '100%',
 	overflowY: 'auto',
-};
-
-export type AssetFileType =
-	| 'audio'
-	| 'video'
-	| 'image'
-	| 'json'
-	| 'txt'
-	| 'other';
-export const getPreviewFileType = (fileName: string | null): AssetFileType => {
-	if (!fileName) {
-		return 'other';
-	}
-
-	const audioExtensions = ['mp3', 'wav', 'ogg', 'aac'];
-	const videoExtensions = ['mp4', 'avi', 'mkv', 'mov', 'webm'];
-	const imageExtensions = ['jpg', 'jpeg', 'png', 'gif', 'bmp'];
-
-	const fileExtension = fileName.split('.').pop()?.toLowerCase();
-	if (fileExtension === undefined) {
-		throw new Error('File extension is undefined');
-	}
-
-	if (audioExtensions.includes(fileExtension)) {
-		return 'audio';
-	}
-
-	if (videoExtensions.includes(fileExtension)) {
-		return 'video';
-	}
-
-	if (imageExtensions.includes(fileExtension)) {
-		return 'image';
-	}
-
-	if (fileExtension === 'json') {
-		return 'json';
-	}
-
-	if (fileExtension === 'txt') {
-		return 'txt';
-	}
-
-	return 'other';
 };
 
 const checkerboardSize = 49;
