+++ packages/studio/src/components/StaticFilePreview.tsx
@@ -3,8 +3,8 @@ import {staticFile} from 'remotion';
 import {StudioServerConnectionCtx} from '../helpers/client-id';
 import {LIGHT_TEXT} from '../helpers/colors';
 import type {AssetMetadata} from '../helpers/get-asset-metadata';
+import {getPreviewFileType} from '../helpers/get-preview-file-type';
 import {FilePreview} from './FilePreview';
-import {getPreviewFileType} from './Preview';
 import {useStaticFiles} from './use-static-files';
 
 const msgStyle: React.CSSProperties = {
