+++ packages/studio/src/helpers/get-preview-file-type.ts
@@ -1,0 +1,44 @@
+export type AssetFileType =
+	| 'audio'
+	| 'video'
+	| 'image'
+	| 'json'
+	| 'txt'
+	| 'other';
+
+export const getPreviewFileType = (fileName: string | null): AssetFileType => {
+	if (!fileName) {
+		return 'other';
+	}
+
+	const audioExtensions = ['mp3', 'wav', 'ogg', 'aac'];
+	const videoExtensions = ['mp4', 'avi', 'mkv', 'mov', 'webm'];
+	const imageExtensions = ['jpg', 'jpeg', 'png', 'gif', 'bmp'];
+
+	const fileExtension = fileName.split('.').pop()?.toLowerCase();
+	if (fileExtension === undefined) {
+		throw new Error('File extension is undefined');
+	}
+
+	if (audioExtensions.includes(fileExtension)) {
+		return 'audio';
+	}
+
+	if (videoExtensions.includes(fileExtension)) {
+		return 'video';
+	}
+
+	if (imageExtensions.includes(fileExtension)) {
+		return 'image';
+	}
+
+	if (fileExtension === 'json') {
+		return 'json';
+	}
+
+	if (fileExtension === 'txt') {
+		return 'txt';
+	}
+
+	return 'other';
+};
